feat : circle progress / lottie loading 분리 적용
