+++ app/src/main/java/com/inha/sellstarter_android/presentation/mypage/MyPageScreen.kt
@@ -7,8 +7,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -16,13 +14,10 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
 import com.inha.sellstarter_android.data.model.request.mypage.UserApiDeleteRequestDto
 import com.inha.sellstarter_android.data.model.request.mypage.UserApiRequestDto
 import com.inha.sellstarter_android.data.model.request.mypage.UserApiUpdateRequest
 import com.inha.sellstarter_android.domain.model.UserInfo
-import com.inha.sellstarter_android.presentation.common.screen.ErrorScreen
-import com.inha.sellstarter_android.presentation.common.screen.LoadingScreen
 import com.inha.sellstarter_android.presentation.common.screen.TitleScreen
 import com.inha.sellstarter_android.presentation.mypage.component.appfont.AppFontSizeContent
 import com.inha.sellstarter_android.presentation.mypage.component.MyPageProfileContent
@@ -31,7 +26,6 @@ import com.inha.sellstarter_android.presentation.mypage.component.appfont.FontSi
 import com.inha.sellstarter_android.presentation.mypage.component.help.HelpContent
 import com.inha.sellstarter_android.presentation.mypage.component.help.helpItems
 import com.inha.sellstarter_android.presentation.mypage.component.storemanage.MyPageStoreAPIContent
-import com.inha.sellstarter_android.util.base.UiState
 
 @Composable
 fun MyPageScreen(
