+++ app/src/main/java/com/inha/sellstarter_android/presentation/order/route/OrderDetailRoute.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.inha.sellstarter_android.presentation.common.screen.ErrorScreen
-import com.inha.sellstarter_android.presentation.common.screen.LoadingScreen
+import com.inha.sellstarter_android.presentation.common.screen.LoadingLottieScreen
 import com.inha.sellstarter_android.presentation.order.detail.OrderDetailScreen
 import com.inha.sellstarter_android.presentation.order.detail.OrderDetailViewModel
 import com.inha.sellstarter_android.util.base.UiState
@@ -30,7 +30,7 @@ fun OrderDetailRoute(
     }
 
     when (val state = detailState) {
-        is UiState.Loading -> LoadingScreen(
+        is UiState.Loading -> LoadingLottieScreen(
             loadingText = "주문 상세 정보를 불러오는 중입니다...",
             modifier = Modifier.fillMaxSize()
         )
